Architecture implemented for components on file setup and last projects updated
