+++ components/home/MainAllies.js
