+++ components/resources/MetricsDataLayout.js
